+++ python/requirements.txt
@@ -1,4 +1,5 @@
 flask
+waitress
 
 python-dotenv
 qdrant_client
